requirements.txt
torch==1.3.1
jieba==0.39
numpy==1.14.5
nltk==3.3
torchvision==0.4.2
tqdm==4.28.1
Pillow==5.3.0
